+++ src/sugar.js
@@ -159,7 +159,7 @@ Observable.prototype.awaiting = function(other) {
 
 Kefir.fromCallback = function(callbackConsumer) {
   var called = false;
-  return Kefir.fromBinder(function(emitter) {
+  return Kefir.stream(function(emitter) {
     if (!called) {
       callbackConsumer(function(x) {
         emitter.emit(x);
@@ -177,7 +177,7 @@ Kefir.fromCallback = function(callbackConsumer) {
 
 Kefir.fromNodeCallback = function(callbackConsumer) {
   var called = false;
-  return Kefir.fromBinder(function(emitter) {
+  return Kefir.stream(function(emitter) {
     if (!called) {
       callbackConsumer(function(error, x) {
         if (error) {
@@ -199,7 +199,7 @@ Kefir.fromNodeCallback = function(callbackConsumer) {
 
 Kefir.fromPromise = function(promise) {
   var called = false;
-  return Kefir.fromBinder(function(emitter) {
+  return Kefir.stream(function(emitter) {
     if (!called) {
       var onValue = function(x) {
         emitter.emit(x);
@@ -229,7 +229,7 @@ Kefir.fromPromise = function(promise) {
 // .fromSubUnsub
 
 function fromSubUnsub(sub, unsub, transformer) {
-  return Kefir.fromBinder(function(emitter) {
+  return Kefir.stream(function(emitter) {
     var handler = transformer ? function() {
       emitter.emit(apply(transformer, this, arguments));
     } : emitter.emit;
@@ -240,7 +240,7 @@ function fromSubUnsub(sub, unsub, transformer) {
   });
 }
 
-Kefir.fromSubUnsub = deprecated('.fromSubUnsub()', '.fromBinder()', fromSubUnsub);
+Kefir.fromSubUnsub = deprecated('.fromSubUnsub()', 'Kefir.stream()', fromSubUnsub);