+++ changelog.md
@@ -1,6 +1,25 @@
+## 2.0.0
+
+### Breaking changes
+
+ - Removed support of old transducers protocol in the `.transduce` (see [#79](https://github.com/pozadi/kefir/issues/79))
+ - `stream.changes()` now returns a new stream with current values/errors removed (see [#56](https://github.com/pozadi/kefir/issues/56))
+ - Properties now can't have both current value and current error at the same time (see [#55](https://github.com/pozadi/kefir/issues/55))
+ - Better errors handling in `.combine` (see [#54](https://github.com/pozadi/kefir/issues/54))
+ - The `.toProperty` method now accepts a callbak instead of a simple value (see [#82](https://github.com/pozadi/kefir/issues/82))
+ - The `.fromEvent` method is renamed to `.fromEvents`
+ - The `.fromBinder` method is renamed to `.stream`
+ - The `.mapEnd` method is renamed to `.beforeEnd` (see [#89](https://github.com/pozadi/kefir/issues/89))
+
+### Other changes
+
+ - The `.fromSubUnsub` method is deprecated (see [#71](https://github.com/pozadi/kefir/issues/71))
+ - Methods `Kefir.emitter()` and `Kefir.bus()` are deprecated (see [#88](https://github.com/pozadi/kefir/issues/88))
+
+
 ## 1.3.1 (2015-04-04)
 
- - The `.transduce` method updated to add support of new protocol [#78](https://github.com/pozadi/kefir/issues/78)
+ - The `.transduce` method updated to add support of new protocol (see [#78](https://github.com/pozadi/kefir/issues/78))
 
 
 ## 1.3.0 (2015-03-29)